Cleanup search handler code
* Move it into its separate file
* Extract filtering of Kibana. This will help with testing later on.
* Introduce better error on invalid version
